Renamed simplecov.rb to simplecov_init.rb to fix require issue
